parse dddml header using `ply`, but nodes using `str.split`
because the header is small but more difficult, so `ply` is appropriate, but the tail can be very large (millions of nodes), but each line is just a list of numbers (separated by spaces), so `ply` is an overkill.

This change yields a speedup by more than an order of magnitude.
The remaining issue is that `networkx` requires a huge amount of memory for half a million nodes.
The root cause are the dictionaries annotating edges.