roots not mandatory any more, nor used anywhere
because reference counting will be used to track both internal and external references. This leads to a simple and uniform interface, instead of using both a set of roots and reference counting.

Moreover, adding the returned edge to the roots would lead to both positive and negative roots, thus having to check for both during garbage collection. To ensure that only positive nodes are added to roots, either the user has to use `abs`, or call a method dedicated to this addition. This does not differ from incrementing a reference counter, using `abs(u)`.

The `BDD.roots` attribute is retained only as a convenience for the user. It is not used anywhere. The user has to explicitly pass the intended roots for each operation (PEP 20: explicit is better than implicit). For example, `to_nx(bdd, bdd.roots)`. This is more flexible, as one can more directly say `to_nx(bed, [u])` w/o the need to change (or care about at all) `bed.roots`.

Also, the restriction that roots should indeed be roots of the digraph is removed from `BDD.assert_consistent`. There is no reason for this: referencing a root function in later operations would invalidate as a root.

Requiring that the intended function node be passed explicitly ensures fewer surprises to the user. For example, `sat_len` and `sat_count` were choosing a root node arbitrarily, which was hidden inside them. This is implicit behavior, to be avoided.