add garbage collection using reference counting
New `BDD` attributes:
- ref
- _min_free
- max_nodes

Explicit reference counting adds a small overhead for users of class `BDD`, but explicit is better than implicit [1]. Regard this as the "low level" interface exposed by `dd.bdd`.

A slight inconvenience is the need to use `abs` on the signed integer representing a returned (possibly complemented) edge. A thin wrapper method may be added for this in the future, but for now things are kept simple (better than complex [1]) and flat (better than nested [1]).

[1] PEP 20