add `BDD` methods: `incref`, `decref`
and hide the attribute `ref`.
This guards against errors of both the users, as well as internally.
It also avoids the bug of decrementing a reference below 0, which prevents garbage collection.
Hiding the attribute allows more flexibility for future changes, as well as experimenting and making available different garbage collection schemes.

Regarding practicality, it is unreasonable to expect users to repeatedly write `bdd.ref[abs(u)] += 1`, and certainly not to fill the place with either selection statements to avoid decrements below 0, or the ugly `bdd.ref[abs(u)] = max(0, bdd.ref[abs(u)])`.

The names `incref` and `decref` originate from the CPython implementation itself [1], namely `Py_INCREF` and `Py_DECREF`.

[1] https://docs.python.org/2/c-api/refcounting.html#reference-counting