add function: `reorder_pairs`
that shifts variables to obtain an order that groups variables into the given pairs. This is substantially more efficient than the brute-force sorting approach.

Even though the number of swaps is about half, the BDD size is <2 times larger, whereas the brute force approach yields a BDD 25 times larger. The reason is that this approach tries to minimally affect the order, remaining closer to the original order (that was (sub)optimized, if the BDD was loaded from a CUDD file with dynamic reordering enables).