add `cython` to `requirements.txt`
because it is auto-detected in `setup.py` and never requested. This complies with the Cython docs recommendation and the time it takes to install Cython. Installing the build dependencies (CUDD or BuDDy), configuring the compiler, and building `dd` extensions will also take some time.

But if Cython is required, even if the above were present, still only `pip` installations would succeed, because `pip` installs dependencies before dependents. Invoking `python setup.py install` would fail, and need to be invoked twice.

Two invocations of `setup.py` are more clumsy than a call to `pip install -r requirements.txt` first. After all, this is a bootstrapping problem, so separate dev environment setup is somewhat justified.