replace `dict.iteritems` with `dd._compat.items(dict)`
to run w/o loss of performance under both Python 2 and 3.
This avoids key iteration with separate value access in the loop body, which is equally efficient, but less readable.

Using `dict.items` is not an option, because `bdd._succ` and other dictionaries contain all the nodes stored, and can be frequently iterated, e.g., during reordering.

Python 2 compatibility is important, because of usage in other packages. In the very distant future, these calls will be replaced with `dict.items`.

Cython is not touched, because it produces C code that can be compiled for any of Python 2 or 3.