ENH: add method `cudd.BDD.sat_iter`
Reuse `bdd._enumerate_minterms`. This requires importing `bdd` in `cudd`, which should better be avoided.
In the future, generic functions like `_enumerate_minterms` can be placed in a common module `dd._utils`.

The function `_enumerate_minterms` is in pure Python, so less efficient.
But this can be ignored, because enumeration itself exhibits the worst-case complexity.