BUG: use `list` of `tuple`s instead of `dict` for parsing renaming pairs
because `astutils.ast.Terminal` is a user-defined class,
but defines `__eq__` and not `__hash__`.
So, it is incorrect to use instances as keys.

Moreover, this is a python 2 weakness.
In python 3, it raises an exception.
The documentation of `object.__hash__` reads [1]:

"A class that overrides __eq__() and
does not define __hash__() will have its __hash__()
implicitly set to None."

[1] https://docs.python.org/3/reference/datamodel.html#object.__hash__