BUG: in Python 2 use `sys.maxint` for `max_nodes`
of `bdd.BDD` and `mdd.MDD`,
because `xrange` in method `BDD._next_free_int` is implemented
using C long (which is limited by `sys.maxint`) and not
`Py_ssize_t`. In Python 3, this limitation does not exist.

Closes #16.

Partially reverts:
    956e15def27ad374263f9c92ef7097d98aa8202b
for the case of Python 2 only.
Should make a difference only on Windows 64 bit, as discussed here:
    https://bugs.python.org/issue26428