BUG: first select level for new var, then check whether occupied
This ensures that no level can be used twice, for two different
variable names, irrespective of how variables are added.
Previously, mixing user-defined and automatically selected levels
could lead to a level associated to an existing variable to
be re-used, which is an error.

This approach raises an `AssertionError` if the user attempts:

```python
import dd

b = dd.bdd.BDD()
b.add_var('b', level=1)  # skip a level
b.add_var('a')  # mix with automated initialization
```

One way to avoid raising an error is to continue scanning for
unoccupied levels, by incrementing the level index.
However, it is avoided, because it nests the decision-making
inside `add_var`. Flat is better than nested (PEP 20).

A situation of this kind can arise only if the user mixes the
two ways of initializing variable levels. So, the user is already
attempting to control the assignment of levels. It is the user's
responsibility to make sure they do so correctly.