API: rm arg `full` from method `BDD.sat_iter`
because it is redundant. The case `full=False` results in
minimal partial assignments as default (those assignments that
result from traversal of the BDD). In absence of `full`,
this case can be obtained for input `care_bits=set()`,
by allowing for `care_bits < support`.

With the current constraint `care_bits >= support`,
and in absence of the argument `full`,
the case `full=True` corresponds to input `care_bits >= support`.

The argument `full` was introduced in commit
9a8b107b1fe46d63df30f86c96a6de838397540e.