API: raise `TypeError` instead of `Exception` if undefined comparison
operator given to `cudd.Function.__richcmp__` (for example `<`).
This allows `pprint` to catch the `TypeError` when it attempts to
compare `Function` objects [1], and finds `__richcmp__` to call for
the purpose of `__lt__` (so Cython itself isn't raising a `TypeError`,
thus execution enters `__richcmp__`).

Apply the same change to `sylvan.Function`.

[1] https://github.com/python/cpython/blob/6a7b3a77b4b2be0badd24ee5f0fdbaa2e0e79c3d/Lib/pprint.py#L89