API: add method `Function.__hash__`
to `autoref, cudd, sylvan` interfaces. Together with the existing
comparator methods, this defines `Function` as hashable.

Previously nodes weren't hashable, even though Python by default
lets user-defined types be hashable, because the presence of
a `__richcmp__` or other comparator method overrides this default
behavior [2].

As the definition of hashable object requires, we implement
`__hash__` so as to ensure that equal `Function` objects have
the same hash value. Nodes that belong to different managers
won't compare equal, so hash conflicts won't cause problems.

That node instances of the same manager with the same pointer (or
same integer node value, in the case of `dd.autoref`) are the same
entity relies on the uniqueness of representation in a BDD.

The only aspects of a node wrapped by a `Function` instance
that can change are the reference count and level of the
corresponding variable. These aren't inherent properties of
the node, and they are the same for all `Function` instances
that wrap that same node, because these values are obtained
from the manager. Thus, BDD nodes are indeed immutable entities.

[1] https://docs.python.org/3/glossary.html
[2] https://docs.python.org/3/reference/datamodel.html#object.__hash__