ENH: add module `dd._abc` as specification
The module `_abc` defines a class `BDD` that documents the API that
other modules implement. The pure Python modules `dd.bdd` and
`dd.autoref` will subclass `BDD` directly from this class, and implement
missing methods (or override as needed). Docstrings are inheritted
in Python 3.5 or higher.

The Cython modules cannot subclass `cdef` classes from `_abc.BDD`.
Instead, a separate script will be used to check whether the
implementation in a Cython module complies to the specification
in `_abc.BDD`. Unfortunately, this leads to duplication of some docstrings.