TST: add script `tests/autoref_test.py` that reuses some `cudd` tests
this is temporary, until an `abc` specification is factored out
of the current classes. Based on the specification, the test suite
will be refactored, to collect reusable tests for the `cudd`,
`autoref`, `sylvan`, `buddy` API in a single script run for all.

The tests that are specific to implementation details of each
module will remain in the module-specific test scripts.

Repetition of some `cudd` tests for `autoref` adds some duplication,
but this duplication is less undesired than minor errors that force
unexpected new releases.