REL: define `long_description` in `setup.py`
instead of using the file `README.md`. The main reason is that
PyPI does not yet recognize Markdown syntax. There are minor
errors when translating from Markdown to reST with `pandoc`,
and such a translation requires more dependencies, which
would complicate the release procedure.

By defining the long description with the file `setup.py`,
an external README file does not need to be opened any more.
The description is a short summary that links to the README
on GitHub for more details.

`numpy` follows a similar approach.