DEP: hide deprecated methods `compose, cofactor, rename`
from module `cudd`. Remove the deprecation warnings from the
hidden methods, because they are called to implement the
cases within the method `BDD.let`.

Rename the hidden method `_compose` to `_unary_compose`,
in order to avoid a naming conflict with the newly hidden
method `_compose`. This choice of name is in accord with
the sibling `_vector_compose`.