REL: create `manylinux1_x86_64` wheel containing CUDD
include in the wheel both the package's `LICENSE` [1] and
CUDD's license as `dd/CUDD_LICENSE`. The latter via
`MANIFEST.in` and `include_package_data=True` for `setuptools.setup`.

CUDD's license is included only in a `bdist_wheel` that
is built with the option `--cudd`. Otherwise this license
is omitted, because no CUDD code is present in that case.

[1] https://github.com/pypa/wheel/blob/c8f23fe728e29b824c1b0bf89913fa02414b0c92/docs/index.rst#including-the-license-in-the-generated-wheel-file