API: implement function `Cudd_bddTransferRename`
in order to enable copying between BDD managers with different
indices for the same variables.

Otherwise, one has to declare variables in the same way for
both managers, which is more complicated, because it requires
calling `cudd.copy_vars`, as opposed to just feeding a list
of variable names to the method `cudd.BDD.declare`.

The function `Cudd_bddTransferRename` is the result of
merging the functions `Cudd_bddTransfer` and `Cudd_bddPermute`,
and their recursive implementations.