MAI: rm inaccurate comment
A level-by-level copying algorithm is as efficient as
multiple calls to an algorithm that copies one root node
in each call, provided the same cache is used for all calls.

Under this assumption about the cache, a level-by-level
copying algorithm performs a breadth-first traversal,
whereas the current implementation of the function `_copy_bdd`
performs a depth-first traversal.