REF: pre-compute levels for unrenamed variables
In preparation of merging the functions `_rename` and
`_copy_bdd`. This change ensures that `dvars` contains
the entire support of `u`.

The entire set of variables is larger than the support.
Nevertheless, restricting the mapping to the support
requires first computing the support, which can require
a traversal of the entire BDD of `u`.

The only faster alternative is to pass a function to
`_rename`, for doing what `dvars.get` was doing.
The reason for this nested approach is to allow
merging `_rename` with `_copy_bdd`.

However, a nested approach is better avoided.
Thus, we pre-compute `dvars`.

Note that in `_copy_bdd` there is no distinction into
"variables to be renamed" and "variables to remain the same".
All variables need to be mapped during `_copy_bdd`,
thus `dict.__getitem__` is used in `_copy_bdd` to guard
against errors.

So before merging `_copy_bdd` and `_rename`, we need to
use `dict.__getitem__` in `_rename` too, instead of
`dict.get` (or pass a callable, which as commented above
is a nested approach). In order to use `dict.__getitem__`,
we have to pre-compute `dvars`, which leads to the change
of `rename` in this commit.