ENH: add method `bdd.BDD.undeclare_vars`
The method `undeclare_vars` removes unused variables (with empty
levels), and renumbers the levels. A subset of unused variables
can be given as arguments, in order to remove fewer variables.

This method modifies variables, levels, and nodes at once.
Thereby the unnecessary swapping of empty variables and repeated
renumbering is avoided.

If no levels are empty, then the full levels can be detected
with an iteration that may terminate before iterating through
all nodes in `self._succ`. Nonetheless, the method `undeclare_vars`
is intended to be called when some level is empty.
So this case is unusual.