BUG: correctly rename variables in `cudd.BDD._rename`
For the call `_rename(u, d)`, if `d['x'] = 'y'` and the BDD `u`
depends on the variable `'y'`, then calling `Cudd_bddSwapVariables`:

1. replaces `'x'` with `'y'` in `u` (as intended)
2. replaces `'y'` with `'x'` in `u` (unintended)

Instead, use composition.

Move the previous implementation of the method `cudd.BDD._rename` to
a new method `cudd.BDD._swap`. The new method asserts that each
variable occurs in at most one key-value pair of the dictionary
`dvars` given as argument.

The implementation of the method `BDD._rename` calls `BDD.var` for
each value of the dictionary `dvars`. The previous implementation
of `BDD._rename` (now the implementation of `BDD._swap`) calls
`BDD.var` for each key and value of `dvars`.

Update the tests of `BDD._rename` to catch this bug.
Test the new method `BDD._swap`.

Comparison to previous changes
==============================

This change partially reverts b3ea0ec6fe27d8900327cd01b128e9e21aee1064,
which removed the assertion that keys and values of `dvars` are
disjoint sets.

The assertion from b3ea0ec6fe27d8900327cd01b128e9e21aee1064 is indeed
restrictive when renaming (simultaneous substitution),
so still absent from `BDD._rename`.

The assertion from b3ea0ec6fe27d8900327cd01b128e9e21aee1064 is required
in `BDD._swap`, in accord with the docstring of the CUDD function
`Cudd_bddSwapVariables` [1].

In addition, it is now asserted that each variable occurs in
at most one key-value pair. This assertion avoids `dvars` values as
for example `dict(x='y', z='y')`.

The assertion from 4ea0e9ae4373ee599163772dec0784c3497a50b3 that the
values of `dvars` are not in the support of the argument `u` is
unnecessary for both `_rename` and `_swap`. However, this assertion
is necessary when swapping is used to implement renaming, and in
that case is restrictive.

This bug was introduced in 4ea0e9ae4373ee599163772dec0784c3497a50b3,
which allowed the target variables to occur in the support of `u`.
This allowed `x /\ y` as `u` and `dvars = dict(x='y')`,
which with swapping results in `x /\ y`. The result of `_rename` in
this case should be `y`. A test that catches this error has been added.

[1] https://github.com/johnyf/cudd/blob/fa6c459cf304d185dd7b573e4dbba84f77070bef/cudd/cuddCompose.c#L416-L417