MAI: reimplement method `cudd.BDD.pick_iter`
The function `pick_iter` was implemented using
CUDD’s generator of cubes. This implementation
appears to yield assignments to variables outside
the support of the BDD node.

The reason for this observation is currently
unknown.

This change reimplements the method `pick_iter`
in Python similarly to how the method
`dd.bdd.BDD.pick_iter` is implemented.