BUG: handle complemented edges correctly in `dd.bdd.BDD.reduction`
Introduced in 0d4b92766d8301fb331163e8e30e51265bd94043,
which changed the representation from plain to complemented edges,
but with the same implementation of the method `BDD.reduction` as
it was written for plain edges [1].

When using complemented edges, the nodes stored as keys in the
dictionary `BDD._succ` are positive integers (not complemented)
[2,3,4,5].

So node `u` is positive [6], thus the dictionary `umap` contains
only positive integers for non-terminal nodes [7] (the key `-1` is
added to `umap` when `umap` is initialized [8]).

As a result, any complemented edge `v` [9] from another level that
points to a non-terminal node is absent from the keys of `umap`,
because it is a negative integer other than `-1`.

Thanks to Sofie Haesaert (@shaesaert) for reporting this error.

[1] https://github.com/tulip-control/dd/commit/f7d3c300292d2d22a7307396e6a583a7948fcc16#diff-523c595745f6d979164fe06ce157b730R195
[2] https://github.com/tulip-control/dd/blob/954c7045ae4a62a50ebc2d5ee063e703cc5307a5/dd/bdd.py#L740
[3] https://github.com/tulip-control/dd/blob/954c7045ae4a62a50ebc2d5ee063e703cc5307a5/dd/bdd.py#L736
[4] https://github.com/tulip-control/dd/blob/954c7045ae4a62a50ebc2d5ee063e703cc5307a5/dd/bdd.py#L742
[5] https://github.com/tulip-control/dd/blob/954c7045ae4a62a50ebc2d5ee063e703cc5307a5/dd/bdd.py#L750
[6] https://github.com/tulip-control/dd/blob/954c7045ae4a62a50ebc2d5ee063e703cc5307a5/dd/bdd.py#L465
[7] https://github.com/tulip-control/dd/blob/954c7045ae4a62a50ebc2d5ee063e703cc5307a5/dd/bdd.py#L468
[8] https://github.com/tulip-control/dd/blob/954c7045ae4a62a50ebc2d5ee063e703cc5307a5/dd/bdd.py#L463
[9] https://github.com/tulip-control/dd/blob/954c7045ae4a62a50ebc2d5ee063e703cc5307a5/dd/bdd.py#L465