REL: bump to `cython == 0.29.15` in `requirements.txt`
`cython < 0.28` expands statements without enclosing
them in parentheses. So the multi-statement C macro
LACE_ME from Sylvan’s `lace.h` works.

In `cython >= 0.28`, Cython encloses the macro
in parentheses as `(LACE_ME)`, which raises
compilation errors, because the macro `LACE_ME`
expands to multiple statements.

The solution was to define an intermediary macro
`LACE_ME_WRAP` as `0); LACE_ME (0`. The fragments
`0);` and `(0` glue the macro to the parentheses
that Cython generates, so that when the macro is
expanded, its multiple statements will not raise
syntax errors.