TST: change assertion in reordering test
Sifting can result in different variable orders,
depending on the order of iteration over
dictionaries and sets in Python, when testing the class
`dd.autoref.BDD`, hence the class `dd.bdd.BDD`.

Instead of asserting that the optimal variable
order is obtained, assert that the size of the BDD
is reduced.