MNT: mv statements outside of `with` statement
The only statements that need to be in the block
of the `with` statement are those that mention the
file `f`. Other statements can be moved outside the
scope of the `with` statement, as is good practice
also for `try...except` statements

Since this `with` statement is equivalent to a
`try...except...finally` statement that always closes
the file (as described below), it is robust compared
to `try...except`. Nonetheless, there is no reason
to increase the size of the `with` statement's body.

In this particular case, [`open` is called in][1] read binary mode,
and buffering is on by default, so [`open` returns][1] an
instance of class `io.BufferedReader`.

The [class `io.BufferedReader`][2] inherits from the
[class `RawIOBase`][3] and the [class `BufferedIOBase`][4],
both of which inherit from the abstract base class
[`io.IOBase`][5]. The class `io.IOBase` is a context manager.
So the open file will be closed in all cases
(for example even if an exception is raised by a statement
inside [the `with` statement's block or target assignment][6]).

[1]: https://docs.python.org/3/library/functions.html#open
[2]: https://docs.python.org/3/library/io.html#io.BufferedReader
[3]: https://docs.python.org/3/library/io.html#io.RawIOBase
[4]: https://docs.python.org/3/library/io.html#io.BufferedIOBase
[5]: https://docs.python.org/3/library/io.html#io.IOBase
[6]: https://docs.python.org/3/reference/compound_stmts.html#with