API: remove the constants `GB`
according to the [ISO/IEC 80000 standard][1], to avoid ambiguity.
The removed constants are `dd.cudd.GB` and `dd.cudd_zdd.GB`.

A "GiB" is a ["gibibyte"][2], prefix ["gibi"][3], and equals
`2**30 bytes`. This is a [binary-multiple prefix][4].

This removal avoids confusion with "GB" according to the
decimal prefix "G" in the ISO standard, which is equal to
[`10**9`][5], and thus ["GB" is equal to `10**9 bytes`][6].

The [JEDEC standard][7] uses "GB" for `2**30`,
which was essentially how "GB" was previously used in the code.

This change is mostly internal: it does not change what `dd`
is doing. It affects only users that directly accessed the
attribute `GB`. Existing code that accessed the attribute `GB`
will now raise an exception (an `AttributeError` or `ImportError`,
depending on how the attribute was accessed).

Raising an error is preferred to other approaches, because:
- it is explicit [PEP 20][9]
- avoids defining unused constants, so simpler [PEP 20][9]
- prevents existing code that relied on the value of `GB` to
  equal `2**30`, from silently running in error due to a change
  in the value of `GB`, which avoids letting errors
  to pass silently [PEP 20][9].

Another approach, without these good characteristics, would be
to add new constants `GiB`, and change the value of existing
`GB` constants to `10**9` (according to ISO).
This would be an implicit change that could go unnoticed,
silently letting dependent code run in error, and would
also introduce unused constants.

The error due to missing `GB` may at first be nonobvious,
but Python introspection and the CHANGELOG can help.

Failing with a message when the module attribute `GB`
is accessed could, in Python 3.7 and later, be approached
based on [PEP 562][8]. However, for now `dd` tries to remain
compatible with earlier Python versions.

Also, update the script `examples/cudd_memory_limits.py`
to not use the constant `GB`.

Mention this change in the CHANGELOG.

[1]: https://en.wikipedia.org/wiki/ISO/IEC_80000#Information_science_and_technology
[2]: https://en.wikipedia.org/wiki/Gibibyte
[3]: https://en.wikipedia.org/wiki/Gibi-
[4]: https://en.wikipedia.org/wiki/Binary_prefix
[5]: https://en.wikipedia.org/wiki/Giga-
[6]: https://en.wikipedia.org/wiki/Gigabyte
[7]: https://en.wikipedia.org/wiki/JEDEC
[8]: https://www.python.org/dev/peps/pep-0562/
[9]: https://www.python.org/dev/peps/pep-0020/