API: return number of bytes from method `statistics`
specifically the value of the key `'mem'` of the `dict`
that is returned from the methods `dd.cudd.BDD.statistics`
and `dd.cudd_zdd.ZDD.statistics`.

This change avoids ambiguity, because ["MB"][1] with ["M"][2]
as in the ISO standard means "`10**6` bytes",
whereas "MB" in the [JEDEC standard][3] means "`2**20` bytes".

Without this change, and for uniformity with the value of the
constant `DEFAULT_MEMORY` in the modules `dd.cudd` and `dd.cudd_zdd`,
the [ISO/IEC 80000 standard][4] prefixes for [binary multiples][5]
would need to be used, together with a division
by `2**20` within the method `statistics`.

This would mean referring to "MiB", i.e.,
["mebibytes"][6] (read also [7][7]).

The need to involve a convention for multipliers is
avoided by returning the number of bytes.

This change can impact users, because the methods
`dd.cudd.BDD.statistics` and `dd.cudd_zdd.ZDD.statistics`
will now return memory values with different meaning.
A [`UserWarning`][8] is now issued, whenever the method
`statistics` is called, to inform the user of this API change.

A consequence of changing what the method `statistics`
returns is the change of what the methods `dd.cudd.BDD.__str__`
and `dd.cudd_zdd.ZDD.__str__` return. This second change
is expected to affect users less.

Mention this change in the CHANGELOG.

[1]: https://en.wikipedia.org/wiki/Megabyte
[2]: https://en.wikipedia.org/wiki/Mega-
[3]: https://en.wikipedia.org/wiki/JEDEC
[4]: https://en.wikipedia.org/wiki/ISO/IEC_80000#Information_science_and_technology
[5]: https://en.wikipedia.org/wiki/Binary_prefix
[6]: https://en.wikipedia.org/wiki/Mebibyte
[7]: https://en.wikipedia.org/wiki/Mebi-
[8]: https://docs.python.org/3/library/exceptions.html#UserWarning