PEP8: uppercase constants in script `setup.py`
The Python variables in module scope that never change
value and are intended to never change value are
constants. Use uppercase letters with underscores for
the identifiers of constants, as recommended in [PEP 8][1].

The script `setup.py` is no exception to this naming
convention. Moreover, most module-scope Python variables
in the script `setup.py` indeed are constants which is
why they are located in module scope.

Some module-scope Python variables in the script `setup.py`
are not constants, for example `networkx_versions`.
These remain named with lowercase letters and underscores.

This discussion evokes considering why these Python variables
are in module scope, instead of inside a function.
Besides being intended to remain constant (being a form of
configuration options), placing them in a function would be
a nested approach (flat is better than nested [PEP 20][2]).

The reduced indentation in module scope is useful for
defining these values, some of which are lengthy.
In addition, placing these values inside the function
`run_setup`, or inside the call to `setuptools.setup`,
would have a denser result (sparse is better than
dense [PEP 20][2]).

[1]: https://www.python.org/dev/peps/pep-0008/#constants
[2]: https://www.python.org/dev/peps/pep-0020/