TST: assert that CUDD node reference count >= 0
Although, for a given BDD/ZDD node, the methods
`dd.cudd.BDD.decref` and `dd.cudd_zdd.ZDD.decref`
were not intended to ever be called without first
calling the corresponding `incref` methods,
this test checks this kind of situation.

A single call to `decref` without a corresponding
earlier call to `incref` was intended to be
possible, but in fact causes errors with the
current implementation, due to the method
`Function.__dealloc__` unconditionally
dereferencing the node.

So such a call to `decref` results, after the call to
`Function.__dealloc__`, in an integer wraparound,
and would have resulted in a *negative* reference count
for the node, had the reference count not been represented
by an unsigned integer within CUDD.

An upcoming commit fixes this error.

Also, the method `decref` is not intended to
be called by the user at all. Instead, automated
reference counting should be used.

In case early release of memory is needed,
then `del u` can be used, where `u` is an instance
of the class `dd.cudd.Function` or of the class
`dd.cudd_zdd.Function`.