MNT: use raw strings where TLA+ backslashes appear
This avoids a `DeprecationWarning` that was introduced
for this syntax in Python 3.6, and is expected to
become a `SyntaxError` in a future Python version.

Quoting from ["Deprecated Python behavior"][1] in "What’s New In Python 3.6":

> A backslash-character pair that is not a valid escape sequence now generates
> a `DeprecationWarning`. Although this will eventually become a `SyntaxError`,
> that will not be for several Python releases.

See also [bpo-27364][3] and [bpo-32912][4].

Quoting from ["String and Bytes literals"][2] in the Python documentation:

> *Changed in version 3.6:* Unrecognized escape sequences produce
> a `DeprecationWarning`. In a future Python version they will be
> a `SyntaxWarning` and eventually a `SyntaxError`.

Revealing these warnings required running with [`python -X dev`][5],
both using `-m pytest`, as well as directly some test scripts.
Note that the setting `filterwarnings = error` in the file
`tests/pytest.ini` (which is introduced in a successor commit)
turns these warnings into errors.

[1]: https://docs.python.org/3/whatsnew/3.6.html#deprecated-python-behavior
[2]: https://docs.python.org/3.10/reference/lexical_analysis.html#string-and-bytes-literals
[3]: https://bugs.python.org/issue27364
[4]: https://bugs.python.org/issue32912
[5]: https://docs.python.org/3/library/devmode.html