API: mv `import pydot` to inline
Even though the package `pydot` is required in the argument
`install_requires` given to the function `setuptools.setup`
in the script `setup.py`, there is no reason to make
the modules `dd.bdd` and `dd.mdd` fail when imported
with an `ImportError` due to absence of `pydot`.

In each of the modules `dd.bdd` and `dd.mdd`, only the
function `to_pydot` uses the package `pydot`. This function
is called only for certain kinds of dumping of a BDD.

No other functionality depends on `pydot`, so the modules
`dd.bdd` and `dd.mdd` can work without `pydot` being installed.
This is made possible by importing `pydot` inside the
function `to_pydot`.