MNT: rm outdated commented code from test script
because if uncommented, then it would raise an error.
The variable `r` points to an instance of the class
`networkx.MultiDiGraph`, which does not have a method
called `write`. The variable `h` points to an instance
of the class `dd.bdd.BDD`, which does not have a method
called `write`.