DOC: add docstring to methods `dd.autoref.BDD.dump` and `load`
because the method `dump` accepts also the file type JSON,
which is not listed in the docstring of the method
`dd._abc.BDD.dump`.

The file type JSON cannot be added to the docstring
of the method `dd._abc.BDD.dump`, because the class
`dd.bdd.BDD` is a subclass of the method `dd._abc.BDD`,
and the method `dd.bdd.BDD.dump` both does not accept
the file type JSON and inherits the docstring of the
method `dd._abc.BDD.dump`.

Similar comments apply to the method `dd.autoref.BDD.load`.