BUG: make argument `filetype` have precedence
over the file extension that may appear in the argument
`filename`, in the method `dd.cudd.BDD.dump`.

In the method `dd.bdd.BDD.dump` and elsewhere, the
parameter `filetype` has precedence over the extension
that appears in the parameter `filename`.

This was not the case in the method `dd.cudd.BDD.dump`.
For example, passing the arguments `filename='foo.dddmp`
and `filetype='json'` would have resulted in writing
a file called `foo.dddmp` with contents in the DDDMP
format.

This change avoids this inconsistency. Even though the
method `dd.cudd.BDD.dump` did not include a docstring,
so the previous behavior of the method was not in conflict
with the method's docstring, this behavior was not intended.

Also, this behavior was in conflict with the docstring
of the method `dd._abc.BDD.dump`, which is intended to
be a specification for both `dd.autoref.BDD.dump` and
`dd.cudd.BDD.dump`, as alluded to in the documentation
file `doc.md`:

> The interface is specified in the module `dd._abc`.