BUG: rm upper bound on `networkx` for all Python versions other than 2
Previously, the conditional statement that selected the
`networkx` version in the script `setup.py` was requiring
that `networkx >= 1.9.1, <= 2.2` also on Python versions
higher than 3.

The place where the supported Python versions are specified
is the parameter `python_requires` of the function
`setuptools.setup` (which is currently not defined in the
script `setup.py`), and in the Trove classifiers.

This approach allows installing this version of `dd` to
future Python versions, if so desired, by suitably
instructing `pip`.