REL: pass parameter `python_requires` to `setuptools.setup`
in the script `setup.py`, to specify the Python versions
that `dd` requires.

The parameter `python_requires` defines the metadata field
`Requires-Python` that is specified in [PEP 345](
    https://www.python.org/dev/peps/pep-0345/#requires-python).
[PEP 503](https://www.python.org/dev/peps/pep-0503/#specification)
specifies how installer tools behave when `Requires-Python` is
present and communicated by PyPI.

Users need to use [`pip >= 9.0.0`](
    https://pip.pypa.io/en/stable/news/#v9-0-0)
and [`setuptools >= 24.3.0`](
    https://setuptools.readthedocs.io/en/latest/history.html#v24-3-0)
for the parameter `python_requires` to have the intended effect.

Relevant resources:
- <https://packaging.python.org/guides/distributing-packages-using-setuptools/#python-requires>
- <https://packaging.python.org/guides/dropping-older-python-versions/>
- <https://packaging.python.org/guides/distributing-packages-using-setuptools/#classifiers>

The documentation in ["Packaging and distributing projects"](
    https://packaging.python.org/guides/distributing-packages-using-setuptools/#classifiers)
of the "Python Packaging User Guide" reads:

> Although the list of classifiers is often used to declare what Python
> versions a project supports, this information is only used for searching &
> browsing projects on PyPI, not for installing projects. To actually
> restrict what Python versions a project can be installed on,
> use the `python_requires` argument.

This seems to answer the speculation below about
what `pip` would do with a package that includes
Trove classifiers for the Python version
and does not have a `pydot_requires`.


`pip install .` uses `python_requires` and
`python setup.py install` ignores `python_requires`.
This behavior was observed with `pip == 21.1.1` and
`setuptools == 56.1.0`.
So `pip` and `setuptools` behave differently,
regarding the parameter `python_requires`.

Both `pip install .` and `python setup.py install`
ignore the Trove classifiers for Python version.
This behavior was observed with
`pip == 21.1.1` and `setuptools == 56.1.0`.

`pip` ignores the Trove classifiers when
`pip install some_sdist.tar.gz`.

`pip` ignores the Trove classifiers also when
`pip install packagename` from PyPI, in case the
package does not specify `Requires-Python` in
its metadata (i.e., in case the parameter
`python_requires` is not passed to the function
`setuptools.setup`).

In contrast, if a package does specify
`Requires-Python`, then `pip` exits
with an error that includes:

> ERROR: Could not find a version that satisfies the requirement

without downloading anything.

Based on these observations, previous releases can be
attempted to be installed on any Python version
(even without any overriding options given to `pip`),
irrespective of the Python versions declared in
the Trove classifiers.

Examples of overriding options are:
`--ignore-requires-python` and `--python-version`.

For these particular versions, this is
less of an issue, because they are compatible
with Python 2.7, which is the use case being considered.