API: `raise AssertionError` instead of `assert` upon deallocation
to ensure that this check is not [removed when optimization][1] is
requested from Python. Even though the exception is raised inside
the method `__dealloc__` or the method `__del__`, thus ignored
by Python (and only a message shown in `sys.stderr`),
this change ensures that at least a message will be shown.

[1]: https://docs.python.org/3/reference/simple_stmts.html#the-assert-statement