MNT: move assertion to safer location within function `_load_json`
Placing the assertion immediately after the call to the
function `_node_from_int` ensures that the node `u` has
just been instantiated, so accessing the attribute `u.ref`
cannot cause access to deallocated memory.

In contrast, after the call to the method `decref`,
it could be the case that the CUDD BDD node that `u` points to
has reached reference count 0. Even though CUDD garbage
collection could not be triggered (even if dynamic reordering
was active, which it is not when `load_order == True`,
which is the only case that the old assertion was nontrivial)
between the call to `decref` and its next line,
it is still better to avoid accessing `u.ref` when the
reference count can be 0.